test-requirements.txt
odoo-addon-base_multicompany_reporting_currency @ git+https://github.com/OCA/sale-reporting.git@refs/pull/223/head#subdirectory=setup/base_multicompany_reporting_currency
